fix(observability): address PR #398 review feedback
Copilot:
- dispatch.rs: replace `format!("{:#04x}", msg_id)` with the raw u8
  field — that heap-allocates on every player-method dispatch and the
  span ships per call when debug is on.
- otel.rs: error message no longer claims spans will ship — the
  function returns None entirely on log-exporter init failure, so
  telemetry is fully off, not partial.
- admin-api/lib.rs: TraceLayer was using tower-http defaults
  (DEBUG/DEBUG/ERROR) but the comment claimed INFO/WARN. Configure
  DefaultMakeSpan + DefaultOnResponse + DefaultOnFailure with the
  promised levels so events flow through the default `info` filter
  without operator tuning.

CodeRabbit:
- tick_sync.rs: per-exit-path disconnect reason. Hardcoded
  "inactivity_timeout" was misclassifying cancelled (logoff) and
  send-error exits. `cancelled` now returns early (caller already
  cleaned up); the timeout and send-error paths break with their
  own labels. Added "send_error" to the documented reason set.
- npc_ai.rs: drop the defensive `_ => {}` wildcard arm — the snapshot
  filter above admits only Fighting | Leashing | Idle, so the
  wildcard is unreachable. Use an if/else if/else chain that
  expresses the snapshot's contract directly (per CLAUDE.md "no
  defensive code for impossible scenarios").
- message_loop.rs + npc_ai.rs: trim the over-narrative comment blocks
  to one-line "why" notes about `.instrument()` vs `.entered()` across
  awaits (per CLAUDE.md "default to no comments").
- helpers.rs (1063 → split into helpers/{mod.rs, tests.rs}): over the
  700-line hard cap (pre-existing). Tests moved to a sibling file;
  mod.rs is now 647 lines.
- encrypted.rs (811 → split into encrypted/{mod.rs, tests.rs}): same
  pattern; mod.rs is now 540 lines.
- login.rs (781 → split into login/{mod.rs, tests.rs}): same pattern;
  mod.rs is now 347 lines.

Codecov:
- Added three tests for the new `disconnect_reason` plumbing through
  `destroy_client_entities`:
  * Single-reason round-trip (the happy path)
  * All five documented reasons round-trip (would catch a refactor
    that drops or renames a label)
  * Idempotent short-circuit path still carries the reason
- All 969 services tests still pass.

Clara: span-coverage expansion suggestions (inventory, gate travel,
AoI events, etc.) deferred — they are scope-expansion follow-ups, not
defects in this PR per Clara's own "strong foundation" assessment.
