+++ crates/services/src/cell/service/npc_ai.rs
@@ -51,36 +51,24 @@ pub(super) async fn npc_ai_tick(tx: &mpsc::Sender<CellToBaseMsg>, space_mgr: &mu
     use tracing::Instrument;
 
     for (npc_id, ai_state, _) in npc_snapshot {
-        // Per-NPC decision span — wraps the handler dispatch so every
-        // log fired inside `npc_ai_fight`/`_leash`/`_idle_auto_aggro`
-        // inherits the npc_id + state correlation. Makes "why isn't
-        // this NPC aggroing me?" a one-trace lookup in SigNoz instead
-        // of grepping debug log lines.
-        //
-        // debug level — only NPCs that passed the eligibility filter
-        // above get here (Fighting/Leashing/aggressive Idle), so the
-        // rate is bounded by population × cadence (2s).
-        //
-        // `.instrument()` (not `.entered()`) — the handler bodies
-        // await, and `.entered()` holds a thread-local guard that
-        // doesn't follow the task across runtime thread switches.
+        // `.instrument()` (not `.entered()`) — the handler bodies await,
+        // so a thread-local guard would silently fall off across runtime
+        // thread switches.
         let ai_span = tracing::debug_span!(
             "npc_ai.decision",
             npc_id,
             ai_state = ?ai_state,
         );
+        // Snapshot filter above admits only Fighting | Leashing | Idle —
+        // express that contract as an if/else if/else rather than a
+        // match-with-wildcard.
         async {
-            match ai_state {
-                AiState::Fighting => {
-                    npc_ai_fight(npc_id, tx, space_mgr).await;
-                }
-                AiState::Leashing => {
-                    npc_ai_leash(npc_id, tx, space_mgr).await;
-                }
-                AiState::Idle => {
-                    npc_ai_idle_auto_aggro(npc_id, tx, space_mgr).await;
-                }
-                _ => {}
+            if ai_state == AiState::Fighting {
+                npc_ai_fight(npc_id, tx, space_mgr).await;
+            } else if ai_state == AiState::Leashing {
+                npc_ai_leash(npc_id, tx, space_mgr).await;
+            } else {
+                npc_ai_idle_auto_aggro(npc_id, tx, space_mgr).await;
             }
         }
         .instrument(ai_span)
